Refactor several classes, add new functionalities and files
The majority of changes involve the renaming and refactoring of several classes in the project. Also, a new file for testing the comparison of records vs classes was added, along with several interfaces and classes in the 'sealed' package. Changes also include a demonstration of the use of break and continue statements. Further adjustments were made to the code in various places for optimisation.